pyproject.toml
[tool.poetry]
# $ pip install pytemplate
name = "diffusion"
version = "0.0.1"
description = "Diffusion models mostly for my own understanding."
authors = ["Bradley"]
license = "LICENSE.txt"

# This field lists other packages that your project depends on to run.
# Any package you put here will be installed by pip when your project is
# installed, so they must be valid existing projects.
[tool.poetry.dependencies]
python = ">=3.9"
matplotlib = "*"
pytest = "*"
pyyaml = "*"
fire = "*"
torch = "*"
torchvision = "*"
tqdm = "*"
pycocotools = "*"

# Linters/formatters
mypy = "*"
pylint = "*"
black = "*"
isort = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"